+++ ntex/CHANGES.md
@@ -4,6 +4,8 @@
 
 * Better names for Variant service
 
+* Add Debug impl for FrozenClientRequest
+
 ## [0.1.24] - 2020-09-22
 
 * Fix ws::stream::StreamDecoder, decodes buffer before reading from io #27